pyproject.toml
[project]
name = "unsloth-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "math-verify>=0.7.0",
    "matplotlib>=3.10.1",
    "pylatexenc>=2.10",
    "scipy>=1.15.2",
    "setuptools>=79.0.0",
    "transformers==4.48.2",
    "trl==0.14.0",
    "unsloth==2025.2.4",
    "vllm==0.7.3",
    "wandb>=0.19.9",
]
